refactor: 유저 도메인 서비스 코드 분리
